+++ types/mempool/sender_nonce.go
@@ -112,12 +112,12 @@ func (snm *SenderNonceMempool) NextSenderTx(sender string) sdk.Tx {
 	}
 
 	cursor := senderIndex.Front()
-	return cursor.Value.(sdk.Tx)
+	return cursor.Value.(MempoolTx).Tx
 }
 
 // Insert adds a tx to the mempool. It returns an error if the tx does not have
 // at least one signer. Note, priority is ignored.
-func (snm *SenderNonceMempool) Insert(_ context.Context, tx sdk.Tx) error {
+func (snm *SenderNonceMempool) InsertWithGasWanted(_ context.Context, tx sdk.Tx, gasLimit uint64) error {
 	snm.mtx.Lock()
 	defer snm.mtx.Unlock()
 	if snm.maxTx > 0 && len(snm.existingTx) >= snm.maxTx {
@@ -126,6 +126,8 @@ func (snm *SenderNonceMempool) Insert(_ context.Context, tx sdk.Tx) error {
 	if snm.maxTx < 0 {
 		return nil
 	}
+
+	memTx := NewMempoolTxWithGasWanted(tx, gasLimit)
 
 	sigs, err := tx.(signing.SigVerifiableTx).GetSignaturesV2()
 	if err != nil {
@@ -145,12 +147,23 @@ func (snm *SenderNonceMempool) Insert(_ context.Context, tx sdk.Tx) error {
 		snm.senders[sender] = senderTxs
 	}
 
-	senderTxs.Set(nonce, tx)
+	senderTxs.Set(nonce, memTx)
 
 	key := txKey{nonce: nonce, address: sender}
 	snm.existingTx[key] = true
 
 	return nil
+}
+
+func (mp *SenderNonceMempool) Insert(ctx context.Context, tx sdk.Tx) error {
+	var gasLimit uint64
+	if gasTx, ok := tx.(interface {
+		GetGas() uint64
+	}); ok {
+		gasLimit = gasTx.GetGas()
+	}
+
+	return mp.InsertWithGasWanted(ctx, tx, gasLimit)
 }
 
 // Select returns an iterator ordering transactions the mempool with the lowest
@@ -268,8 +281,8 @@ func (i *senderNonceMempoolIterator) Next() Iterator {
 	return nil
 }
 
-func (i *senderNonceMempoolIterator) Tx() sdk.Tx {
-	return i.currentTx.Value.(sdk.Tx)
+func (i *senderNonceMempoolIterator) Tx() MempoolTx {
+	return i.currentTx.Value.(MempoolTx)
 }
 
 func removeAtIndex[T any](slice []T, index int) []T {